Revert "Chore/sign up"
